Add version string to lldb-vscode
Summary:
There is a frequent need to know the version of lldb-vscode VSCode is using during scuba query so that we can determine if a feature/bug fix is in the query set or not.

This patch enables this by adding version string to lldb-vscode `initialized` event so that VScode telemetry can record in scuba.

Test Plan: See version string in DAP log.

Reviewers: hyubo, #lldb_team

Reviewed By: hyubo

Subscribers: #lldb_team

Differential Revision: https://phabricator.intern.facebook.com/D42103090